remove _app.js, make index
